Re-read loop device's partition table before pv
Currently the scripts/chromeos-install.sh install script doesn't re-read
loop device's partition table before copying the partitions to the usb
drive. This causes pv not to find the partitions if the loop device was
used with different image before.

To fix that, partprobe the loop device before pv-ing its partitions to
the target device.

Fixes #97